properly generating extended tests -- need to get nesting to print properly and have the right IDs

+++ df-testgen/src/main.rs
@@ -42,16 +42,22 @@ fn setup_toy_fs(path_start: &str) -> Result<Vec<PathBuf>, std::io::Error> {
 
     for dir in &decisions::setup::TOY_FS_DIRS {
         let cur_path = PathBuf::from(path_start.to_owned() + "/" + dir);
+        toy_fs_paths.push(cur_path.clone());
+        if Path::new(&(cur_path)).exists() {
+            continue;
+        }
         std::fs::DirBuilder::new()
             .recursive(true)
             .create(&cur_path)?;
-        toy_fs_paths.push(cur_path);
     }
 
     for file in &decisions::setup::TOY_FS_FILES {
         let cur_path = PathBuf::from(path_start.to_owned() + "/" + file);
+        toy_fs_paths.push(cur_path.clone());
+        if Path::new(&(cur_path)).exists() {
+            continue;
+        }
         std::fs::File::create(path_start.to_owned() + "/" + file)?;
-        toy_fs_paths.push(cur_path);
     }
 
     Ok(toy_fs_paths)
